+++ sensors/Pressure.h
@@ -4,4 +4,4 @@
 #include <stdbool.h>
 
 bool Calibrate_Sensor(uint16_t diff);
-float Convert_Pressure(uint16_t diff);
+float Convert_Pressure(float diff);
